Lots of QoL changes
moved macro section to top of lib.rs.  This resolves a lot of the
importing of macro issues.

Fixed all compiler warnings.  This makes it easier to find compiler
errors.

Currently Staticvec is disabled.  I renamed the file to
staticvec.rs.unused just so we don't get all of the warning messages
about it.  Also the mod line for it in /utils/mod.rs is commented
out so if you need it in the future you will need to fix this as well.